@@ -2,7 +2,6 @@
   "name": "vite-react-express-boilerplate",
   "version": "0.0.0",
   "scripts": {
-    "clean": "rimraf dist",
     "dev": "concurrently \"npm run server:dev\" \"npm run client:dev\"",
     "client:dev": "vite",
     "server:dev": "nodemon src/server.js",
@@ -10,38 +9,36 @@
     "client:build": "vite build",
     "build": "npm run server:build && npm run client:build",
     "serve": "vite preview",
-    "prepare": "husky install",
-    "test": "prettier --check . && eslint . && stylelint \"**/*.css\"",
     "start": "node dist/server.js"
   },
   "devDependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.7",
-    "@vitejs/plugin-react": "^1.0.2",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@vitejs/plugin-react": "^1.3.1",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "concurrently": "^6.2.1",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "http-proxy-middleware": "^2.0.1",
-    "husky": "^7.0.2",
-    "lint-staged": "^11.1.2",
+    "eslint": "^8.13.0",
+    "eslint-config-eda": "^1.1.0",    "http-proxy-middleware": "^2.0.1",
     "nodemon": "^2.0.13",
-    "prettier": "2.4.1",
     "prop-types": "^15.7.2",
-    "rimraf": "^3.0.2",
-    "stylelint": "^13.13.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-standard": "^22.0.0",
+    "react-redux": "^8.0.1",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
     "vite": "^2.5.10"
   },
   "dependencies": {
-    "dotenv": "^10.0.0",
+    "@chakra-ui/react": "^2.0.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "dotenv": "^16.0.0",
     "express": "^4.17.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "framer-motion": "^6.3.3",
+    "knex": "^1.0.7",
+    "react": "^18.1.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
+    "superagent": "^7.1.2"
   }
 }
